requirements.txt
PyQt6
openai-whisper
torch
openai
